Only allow dragging from top part of window
`QMainWindow` allows dragging any empty part of the window to move
it, as well as double clicking to maximize/minimize.
This patch adds to the existing event filter for `frame`, listening
for `MouseButtonPress` and `MouseButtonDblClick` events, only
allowing them through if the click's Y coordinate is anywhere above
the search text edit field (`ui->searchEdit`).

Fixes #173