Travis CI: Removed the failing 'trusty' build
Since the addition of qmarkdowntextedit, APIs are used that were added
in Qt 5.3. Unless we're interested in restoring compatibility with
Qt 5.2 (probably possible by changing qmarkdowntextedit), I would suggest
we simply drop this Trusty build, who's end of support was reached in
April 2019.

I've also added a Qt version check, along with a Qt 5.3 build to check
whether Notes really does compile against that Qt version. Of course, we
may want to raise the minimum Qt version further, say to Qt 5.5, which
is the version in the current oldest supported Ubuntu (xenial).